+++ web-app/src/app/screens/Feeds/Feeds.styles.ts
@@ -1,8 +1,11 @@
 import { styled, type SxProps, type Theme, Typography } from '@mui/material';
 
 export const SearchHeader = styled(Typography)(({ theme }) => ({
+  '&.no-collapse': {
+    margin: '12px 0',
+  },
   '&:not(:first-of-type)': {
-    marginTop: theme.spacing(2),
+    marginTop: theme.spacing(1),
   },
   '&:after': {
     content: '""',
