+++ web-app/src/app/screens/Feeds/index.tsx
@@ -29,10 +29,6 @@ import {
 import { useSearchParams } from 'react-router-dom';
 import SearchTable from './SearchTable';
 import { Trans, useTranslation } from 'react-i18next';
-import { DATASET_FEATURES, groupFeaturesByComponent } from '../../utils/consts';
-import NestedCheckboxList, {
-  type CheckboxStructure,
-} from '../../components/NestedCheckboxList';
 import {
   getDataTypeParamFromSelectedFeedTypes,
   getInitialSelectedFeedTypes,
@@ -40,7 +36,6 @@ import {
 import {
   chipHolderStyles,
   searchBarStyles,
-  SearchHeader,
   stickyHeaderStyles,
 } from './Feeds.styles';
 import { useRemoteConfig } from '../../context/RemoteConfigProvider';
@@ -49,6 +44,7 @@ import { ColoredContainer } from '../../styles/PageLayout.style';
 import AdvancedSearchTable from './AdvancedSearchTable';
 import ViewHeadlineIcon from '@mui/icons-material/ViewHeadline';
 import GridViewIcon from '@mui/icons-material/GridView';
+import { SearchFilters } from './SearchFilters';
 
 export default function Feed(): React.ReactElement {
   const theme = useTheme();
@@ -65,18 +61,13 @@ export default function Feed(): React.ReactElement {
     Boolean(searchParams.get('official')) ?? false,
   );
   const [searchQuery, setSearchQuery] = useState(activeSearch);
-  const [expandedElements, setExpandedElements] = useState<
-    Record<string, boolean>
-  >(setInitialExpandGroup());
+
   const [selectedFeatures, setSelectedFeatures] = useState<string[]>(
     searchParams.get('features')?.split(',') ?? [],
   );
   const [selectGbfsVersions, setSelectGbfsVersions] = useState<string[]>(
     searchParams.get('gbfs_versions')?.split(',') ?? [],
   );
-  const [featureCheckboxData, setFeatureCheckboxData] = useState<
-    CheckboxStructure[]
-  >([]);
   const [activePagination, setActivePagination] = useState(
     searchParams.get('o') !== null ? Number(searchParams.get('o')) : 1,
   );
@@ -134,7 +125,7 @@ export default function Feed(): React.ReactElement {
             status: ['active', 'inactive', 'development', 'future'],
             feature: areFeatureFiltersEnabled ? selectedFeatures : undefined,
             version: areGBFSFiltersEnabled
-              ? selectGbfsVersions.join(',')
+              ? selectGbfsVersions.join(',').replaceAll('v', '')
               : undefined,
           },
         },
@@ -256,43 +247,6 @@ export default function Feed(): React.ReactElement {
     }
   };
 
-  function setInitialExpandGroup(): Record<string, boolean> {
-    const expandGroup: Record<string, boolean> = {};
-    Object.keys(
-      groupFeaturesByComponent(Object.keys(DATASET_FEATURES), true),
-    ).forEach((featureGroup) => {
-      expandGroup[featureGroup] = false;
-    });
-    return expandGroup;
-  }
-
-  function generateCheckboxStructure(): CheckboxStructure[] {
-    const groupedFeatures = groupFeaturesByComponent(
-      Object.keys(DATASET_FEATURES),
-      true,
-    );
-    return Object.entries(groupedFeatures)
-      .filter(([parent]) => parent !== 'Other')
-      .sort(([keyA], [keyB]) => keyA.localeCompare(keyB))
-      .map(([parent, features]) => ({
-        title: parent,
-        checked: features.every((feature) =>
-          selectedFeatures.includes(feature.feature),
-        ),
-        seeChildren: expandedElements[parent],
-        type: 'checkbox',
-        children: features.map((feature) => {
-          return {
-            title: feature.feature,
-            type: 'checkbox',
-            checked: selectedFeatures.some(
-              (selectedFeature) => selectedFeature === feature.feature,
-            ),
-          };
-        }),
-      }));
-  }
-
   function clearAllFilters(): void {
     setActivePagination(1);
     setSelectedFeedTypes({
@@ -304,10 +258,6 @@ export default function Feed(): React.ReactElement {
     setSelectGbfsVersions([]);
     setIsOfficialFeedSearch(false);
   }
-
-  React.useEffect(() => {
-    setFeatureCheckboxData(generateCheckboxStructure());
-  }, [selectedFeatures]);
 
   const containerRef = React.useRef(null);
   useEffect(() => {
@@ -335,26 +285,6 @@ export default function Feed(): React.ReactElement {
       setSearchView(newSearchView);
     }
   };
-  const dataTypesCheckboxData: CheckboxStructure[] = [
-    {
-      title: t('common:gtfsSchedule'),
-      checked: selectedFeedTypes.gtfs,
-      type: 'checkbox',
-    },
-    {
-      title: t('common:gtfsRealtime'),
-      checked: selectedFeedTypes.gtfs_rt,
-      type: 'checkbox',
-    },
-  ];
-
-  if (config.enableGbfsInSearchPage) {
-    dataTypesCheckboxData.push({
-      title: t('common:gbfs'),
-      checked: selectedFeedTypes.gbfs,
-      type: 'checkbox',
-    });
-  }
 
   return (
     <Container
@@ -452,128 +382,31 @@ export default function Feed(): React.ReactElement {
                 pr: 2,
               }}
             >
-              <SearchHeader variant='h6'>{t('dataType')}</SearchHeader>
-              <NestedCheckboxList
-                checkboxData={dataTypesCheckboxData}
-                onCheckboxChange={(checkboxData) => {
+              <SearchFilters
+                selectedFeedTypes={selectedFeedTypes}
+                isOfficialFeedSearch={isOfficialFeedSearch}
+                selectedFeatures={selectedFeatures}
+                selectedGbfsVersions={selectGbfsVersions}
+                setSelectedFeedTypes={(feedTypes) => {
                   setActivePagination(1);
-                  const checkedFeedTypes = {
-                    ...selectedFeedTypes,
-                    gtfs: checkboxData[0].checked,
-                    gtfs_rt: checkboxData[1].checked,
-                    gbfs: false,
-                  };
-                  if (config.enableGbfsInSearchPage) {
-                    checkedFeedTypes.gbfs = checkboxData[2].checked;
-                  }
-                  setSelectedFeedTypes(checkedFeedTypes);
+                  setSelectedFeedTypes(feedTypes);
                 }}
-              ></NestedCheckboxList>
-              {config.enableIsOfficialFilterSearch && (
-                <>
-                  <SearchHeader
-                    variant='h6'
-                    sx={isOfficialTagFilterEnabled ? {} : { opacity: 0.5 }}
-                  >
-                    Tags
-                  </SearchHeader>
-                  <NestedCheckboxList
-                    disableAll={!isOfficialTagFilterEnabled}
-                    checkboxData={[
-                      {
-                        title: 'Official Feeds',
-                        checked: isOfficialFeedSearch,
-                        type: 'checkbox',
-                      },
-                    ]}
-                    onCheckboxChange={(checkboxData) => {
-                      setActivePagination(1);
-                      setIsOfficialFeedSearch(checkboxData[0].checked);
-                    }}
-                  ></NestedCheckboxList>
-                </>
-              )}
-
-              {config.enableFeatureFilterSearch && (
-                <>
-                  <SearchHeader
-                    variant='h6'
-                    sx={areFeatureFiltersEnabled ? {} : { opacity: 0.5 }}
-                  >
-                    Features
-                  </SearchHeader>
-                  <NestedCheckboxList
-                    disableAll={!areFeatureFiltersEnabled}
-                    debounceTime={500}
-                    checkboxData={featureCheckboxData}
-                    onExpandGroupChange={(checkboxData) => {
-                      const newExpandGroup: Record<string, boolean> = {};
-                      checkboxData.forEach((cd) => {
-                        if (cd.seeChildren !== undefined) {
-                          newExpandGroup[cd.title] = cd.seeChildren;
-                        }
-                      });
-                      setExpandedElements({
-                        ...expandedElements,
-                        ...newExpandGroup,
-                      });
-                    }}
-                    onCheckboxChange={(checkboxData) => {
-                      const selelectedFeatures: string[] = [];
-                      checkboxData.forEach((checkbox) => {
-                        if (checkbox.children !== undefined) {
-                          checkbox.children.forEach((child) => {
-                            if (child.checked) {
-                              selelectedFeatures.push(child.title);
-                            }
-                          });
-                        }
-                      });
-                      setActivePagination(1);
-                      setSelectedFeatures([...selelectedFeatures]);
-                    }}
-                  />
-                </>
-              )}
-
-              {config.enableGbfsInSearchPage && (
-                <>
-                  <SearchHeader
-                    variant='h6'
-                    sx={areGBFSFiltersEnabled ? {} : { opacity: 0.5 }}
-                  >
-                    GBFS Versions
-                  </SearchHeader>
-                  <NestedCheckboxList
-                    disableAll={!areGBFSFiltersEnabled}
-                    debounceTime={500}
-                    checkboxData={[
-                      '3.1-RC',
-                      '3.0',
-                      '2.3',
-                      '2.2',
-                      '2.1',
-                      '2.0',
-                      '1.1',
-                      '1.0',
-                    ].map((version) => ({
-                      title: version,
-                      checked: selectGbfsVersions.includes(version),
-                      type: 'checkbox',
-                    }))}
-                    onCheckboxChange={(checkboxData) => {
-                      const selectedVersions: string[] = [];
-                      checkboxData.forEach((checkbox) => {
-                        if (checkbox.checked) {
-                          selectedVersions.push(checkbox.title);
-                        }
-                      });
-                      setActivePagination(1);
-                      setSelectGbfsVersions([...selectedVersions]);
-                    }}
-                  ></NestedCheckboxList>
-                </>
-              )}
+                setIsOfficialFeedSearch={(isOfficial) => {
+                  setActivePagination(1);
+                  setIsOfficialFeedSearch(isOfficial);
+                }}
+                setSelectedFeatures={(features) => {
+                  setActivePagination(1);
+                  setSelectedFeatures(features);
+                }}
+                setSelectedGbfsVerions={(versions) => {
+                  setSelectGbfsVersions(versions);
+                  setActivePagination(1);
+                }}
+                isOfficialTagFilterEnabled={isOfficialTagFilterEnabled}
+                areFeatureFiltersEnabled={areFeatureFiltersEnabled}
+                areGBFSFiltersEnabled={areGBFSFiltersEnabled}
+              ></SearchFilters>
             </Grid>
 
             <Grid item xs={12} md={10}>
@@ -673,7 +506,8 @@ export default function Feed(): React.ReactElement {
                   selectGbfsVersions.length > 0 ||
                   isOfficialFeedSearch ||
                   selectedFeedTypes.gtfs_rt ||
-                  selectedFeedTypes.gtfs || selectedFeedTypes.gbfs) && (
+                  selectedFeedTypes.gtfs ||
+                  selectedFeedTypes.gbfs) && (
                   <Button
                     variant={'text'}
                     onClick={clearAllFilters}
