+++ web-app/src/app/utils/gbfsConsts.ts
@@ -1,0 +1,10 @@
+export const gbfsVersions = [
+  'v3.1-RC',
+  'v3.0',
+  'v2.3',
+  'v2.2',
+  'v2.1',
+  'v2.0',
+  'v1.1',
+  'v1.0',
+];
